requirements.txt
absl-py
matplotlib
seaborn
umap-learn
graphgrove
numpy
tqdm

# For imagenet download
six
tensorflow==2.13.0
kaggle